Refactor setUp and tearDown method to Skeleton Command Base Test
Test methods for Generated files are extracted into
SkeletonGeneratedFilesTest
